feat: add tikl submodule and config

+++ test/integration/tsano_inflex_abc.c
@@ -2,15 +2,13 @@
 // UNSUPPORTED: aarch64
 // RUN: (! %lotto %stress --stable-address-method MASK -- %b 2>&1) | %check %s --check-prefix=BUG
 // RUN: %lotto %inflex -r 50 &>/dev/null
-// RUN: %lotto %debug --file-filter="libvsync" <<< $'\n'run-replay-lotto | %check %s --check-prefix=LOC
+// RUN: %lotto %debug <<< $'\n'run-replay-lotto | %check %s --check-prefix=LOC
 // BUG: assert failed {{.*}}/tsano_inflex_abc.c:{{[0-9]+}}: recv_from[i] == true
 // LOC: uint32_t id = vatomic32_read(&next_id);
 // clang-format on
 #include <assert.h>
 #include <pthread.h>
 #include <stdbool.h>
-#include <stdio.h>
-#include <stdlib.h>
 
 #include <vsync/queue/bounded_mpmc.h>
 
@@ -43,7 +41,7 @@ producer(void *_)
     (void)_;
 
     int *data = (int *)_; // malloc(sizeof(int));
-    *data     = get_unique_id();
+    *data     = (int)get_unique_id();
     while (1) {
         int r = bounded_mpmc_enq(&q, data);
         if (r == QUEUE_BOUNDED_OK)
